trial & complete about EncryptionOperationNotPossibleException Error
