+++ .settings/org.eclipse.wst.common.component
@@ -1,17 +1,19 @@
 <?xml version="1.0" encoding="UTF-8"?><project-modules id="moduleCoreId" project-version="1.5.0">
-        
-    <wb-module deploy-name="WEBSITE">
-                
-        <wb-resource deploy-path="/" source-path="/src/main/webapp" tag="defaultRootSource"/>
-                
-        <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/resources"/>
-                
-        <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/test/resources"/>
-                
-        <property name="context-root" value="WEBSITE"/>
-                
-        <property name="java-output-path" value="/WEBSITE/target/classes"/>
             
-    </wb-module>
     
+    <wb-module deploy-name="WEBSITE">
+        <wb-resource deploy-path="/" source-path="/src/main/webapp" tag="defaultRootSource"/>
+        <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/resources"/>
+        <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/java"/>
+                        
+        
+        <property name="context-root" value="WEBSITE"/>
+                        
+        
+        <property name="java-output-path" value="/WEBSITE/target/classes"/>
+                    
+    
+    </wb-module>
+        
+
 </project-modules>
