+++ src/main/resources/config.properties
@@ -1,3 +1,3 @@
 db.url=jdbc:mysql://localhost:3306/WEB_SECURITY_LAB
 db.user=root
-db.password=ENC(vzHXBj4Lys1g3Vu//uPXiA==)
+db.password=ENC(SBgHo9N3OIWlY2/ql2jawg==)
